third_party/WebKit/LayoutTests/http/tests/credentialmanager/credentialscontainer-notifyfailedsignin-basics.html

<!DOCTYPE html>
<title>Credential Manager: notifyFailedSignIn() basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
function stubResolverChecker(c) {
    assert_equals(c, undefined, "FIXME: We're currently always returning 'undefined'.");
    this.done();
}

function stubRejectionChecker(reason) {
    assert_unreached("notifyFailedSignIn() should not reject, but did: " + reason);
}

var local = new PasswordCredential('id', 'pencil', 'name', 'https://example.com/avatar.png');
var federated = new FederatedCredential({
    'id': 'id',
    'provider': 'https://federation.test/',
    'name': 'name',
    'iconURL': 'https://example.test/icon.png'
});

async_test(function () {
    navigator.credentials.notifyFailedSignIn().then(
        this.step_func(function () { assert_unreached("notifyFailedSignIn() should reject."); }),
        this.step_func(function (reason) {
            assert_equals(reason.name, "TypeError");
            this.done();
        }));
}, "Verify the basics of notifyFailedSignIn(): PasswordCredential.");

async_test(function () {
    navigator.credentials.notifyFailedSignIn(local).then(
        this.step_func(stubResolverChecker.bind(this)),
        this.step_func(stubRejectionChecker.bind(this)));
}, "Verify the basics of notifyFailedSignIn(): PasswordCredential.");

async_test(function () {
    navigator.credentials.notifyFailedSignIn(federated).then(
        this.step_func(stubResolverChecker.bind(this)),
        this.step_func(stubRejectionChecker.bind(this)));
}, "Verify the basics of notifyFailedSignIn(): FederatedCredential.");
</script>
